mail update (all run in series)

+++ costantini_code/parameters_setting.py
@@ -6,9 +6,7 @@ def changing_parameters1(parameters, table, sector=4, layer=0, component=0):
     if isinstance(table, pd.DataFrame):
         print("dentro primo if")
         if isinstance(parameters, list):
-            table['sector']=table['sector'].astype(int)
-            table['layer']=table['layer'].astype(int)
-            table['component']=table['component'].astype(int)
+
 
             table.loc[((table['sector'] == sector) & (table['layer'] == layer) & (
                         table['component'] == component)), table.columns.tolist()] = [sector, layer,
@@ -33,6 +31,10 @@ def changing_parameters(parameters, table, module=None):
 
     if isinstance(table, pd.DataFrame):
         if isinstance(parameters, list):
+            table['sector'] = table['sector'].astype(int)
+            table['layer'] = table['layer'].astype(int)
+            table['component'] = table['component'].astype(int)
+
             table.loc[((table['sector'] == sector) & (table['layer'] == layer) & (
                         table['component'] == component)), table.columns.tolist()] = [sector, layer,
                                                                                       component] + parameters
